feat(NormalizedCache): use a Map-like cache abstraction
This feature enables a swappable cache implementation. For example, you might want to use Map, which is faster than writing keys to an Object. This also allows for custom use cases, such as emitting events upon .set() or .delete() (think Observables), which was otherwise impossible without the use of Proxies, which were only available in ES6.